distribute blue squad tests to libtest and workload folders
