+++ tests/functional/monitoring/workload/test_workload_with_distruptions.py
@@ -33,12 +33,14 @@ import logging
 import time
 import pytest
 from ocs_ci.framework import config
-from ocs_ci.framework.pytest_customization.marks import blue_squad, tier3
 from ocs_ci.framework.testlib import (
+    blue_squad,
+    tier1,
     tier2,
     pre_upgrade,
     post_upgrade,
     skipif_managed_service,
+    skipif_mcg_only,
     skipif_hci_provider_and_client,
 )
 from ocs_ci.ocs import constants, ocp
@@ -51,9 +53,6 @@ from ocs_ci.ocs.cluster import (
 )
 from ocs_ci.ocs.fiojob import get_timeout
 from ocs_ci.ocs.resources.objectconfigfile import ObjectConfFile
-from ocs_ci.ocs.resources.pod import get_osd_pods
-from ocs_ci.utility import prometheus
-from ocs_ci.utility.prometheus import PrometheusAPI
 
 logger = logging.getLogger(__name__)
 
@@ -189,149 +188,41 @@ def test_workload_with_checksum_verify(
     logger.info("sha1sum output: %s", sha1sum_output)
 
 
-class TestCephOSDSlowOps(object):
-    @pytest.fixture(scope="function")
-    def setup(self, request, pod_factory, multi_pvc_factory):
-        """
-        Set preconditions to trigger CephOSDSlowOps
-        """
-        self.test_pass = None
-        reduced_osd_complaint_time = 0.1
+@blue_squad
+@skipif_mcg_only
+@tier1
+@pytest.mark.polarion_id("OCS-2125")
+@skipif_managed_service
+def test_workload_rbd_cephfs_10g(
+    workload_storageutilization_10g_rbd, workload_storageutilization_10g_cephfs
+):
+    """
+    Test of a workload utilization with constant 10 GiB target.
 
-        set_osd_op_complaint_time(reduced_osd_complaint_time)
+    In this test we are only checking whether the storage utilization workload
+    failed or not. The main point of having this included in tier1 suite is to
+    see whether we are able to actually run the fio write workload without any
+    direct failure (fio job could fail to be scheduled, fail during writing or
+    timeout when write progress is too slow ...).
+    """
+    logger.info("checking fio report results as provided by workload fixtures")
+    msg = "workload results should be recorded and provided to the test"
+    assert workload_storageutilization_10g_rbd["result"] is not None, msg
+    assert workload_storageutilization_10g_cephfs["result"] is not None, msg
 
-        ceph_cluster = CephCluster()
-
-        self.full_osd_ratio = round(get_full_ratio_from_osd_dump(), 2)
-        self.full_osd_threshold = self.full_osd_ratio * 100
-
-        # max possible cap to reach CephOSDSlowOps is to fill storage up to threshold; alert should appear much earlier
-        pvc_size = ceph_cluster.get_ceph_free_capacity() * self.full_osd_ratio
-
-        # assuming storageutilization speed reduced to less than 1, estimation timeout to fill the storage
-        # will be reduced by number of osds. That should be more than enough to trigger an alert,
-        # otherwise the failure is legitimate
-        storageutilization_min_mbps = config.ENV_DATA[
-            "fio_storageutilization_min_mbps"
-        ] / len(get_osd_pods())
-        self.timeout_sec = get_timeout(storageutilization_min_mbps, int(pvc_size))
-
-        access_modes = [constants.ACCESS_MODE_RWO, constants.ACCESS_MODE_RWX]
-
-        num_of_load_objs = 2
-        self.pvc_objs = multi_pvc_factory(
-            interface=constants.CEPHFILESYSTEM,
-            size=pvc_size / num_of_load_objs,
-            access_modes=access_modes,
-            status=constants.STATUS_BOUND,
-            num_of_pvc=num_of_load_objs,
-            wait_each=True,
+    fio_reports = (
+        ("rbd", workload_storageutilization_10g_rbd["result"]["fio"]),
+        ("cephfs", workload_storageutilization_10g_cephfs["result"]["fio"]),
+    )
+    for vol_type, fio in fio_reports:
+        logger.info("starting to check fio run on %s volume", vol_type)
+        msg = "single fio job should be executed in each workload run"
+        assert len(fio["jobs"]) == 1, msg
+        logger.info(
+            "fio (version %s) executed %s job on %s volume",
+            fio["fio version"],
+            fio["jobs"][0]["jobname"],
+            vol_type,
         )
-        self.pod_objs = []
-
-        for pvc_obj in self.pvc_objs:
-            pod_obj = pod_factory(
-                interface=constants.CEPHFILESYSTEM, pvc=pvc_obj, replica_count=3
-            )
-            self.pod_objs.append(pod_obj)
-            file_name = pod_obj.name
-            pod_obj.fillup_fs(size=f"{round(pvc_size * 1024)}M", fio_filename=file_name)
-            pod_obj.run_io(
-                storage_type="fs",
-                size="3G",
-                runtime=self.timeout_sec,
-                fio_filename=f"{pod_obj.name}_io",
-            )
-
-        self.start_workload_time = time.perf_counter()
-
-        def finalizer():
-            """
-            Set default values for:
-              osd_op_complaint_time=30.000000
-            """
-            # set the osd_op_complaint_time to selected monitor back to default value
-            set_osd_op_complaint_time(constants.DEFAULT_OSD_OP_COMPLAINT_TIME)
-
-            # delete resources
-            for pod_obj in self.pod_objs:
-                pod_obj.delete()
-                pod_obj.delete(wait=True)
-
-            for pvc_obj in self.pvc_objs:
-                pvc_obj.delete(wait=True)
-
-        request.addfinalizer(finalizer)
-
-    @tier3
-    @pytest.mark.polarion_id("OCS-5158")
-    @blue_squad
-    def test_ceph_osd_slow_ops_alert(self, setup, threading_lock):
-        """
-        Test to verify bz #1966139, more info about Prometheus alert - #1885441
-
-        CephOSDSlowOps. An Object Storage Device (OSD) with slow requests is every OSD that is not able to service
-        the I/O operations per second (IOPS) in the queue within the time defined by the osd_op_complaint_time
-        parameter. By default, this parameter is set to 30 seconds.
-
-        1. As precondition test setup is to reduce osd_op_complaint_time to 0.1 to prepare condition
-        to get CephOSDSlowOps
-        2. Run workload_fio_storageutilization gradually filling up the storage up to full_ratio % in a background
-        2.1 Validate the CephOSDSlowOps fired, if so check an alert message and finish the test
-        2.2 If CephOSDSlowOps has not been fired while the storage filled up to full_ratio % or time to fill up the
-        storage ends - fail the test
-        """
-
-        api = PrometheusAPI(threading_lock=threading_lock)
-
-        while get_percent_used_capacity() < self.full_osd_threshold:
-            time_passed_sec = time.perf_counter() - self.start_workload_time
-            if time_passed_sec > self.timeout_sec:
-                pytest.fail("failed to fill the storage in calculated time")
-
-            delay_time = 60
-            logger.info(f"sleep {delay_time}s")
-            time.sleep(delay_time)
-
-            alerts_response = api.get(
-                "alerts", payload={"silenced": False, "inhibited": False}
-            )
-            if not alerts_response.ok:
-                logger.error(
-                    f"got bad response from Prometheus: {alerts_response.text}"
-                )
-                continue
-            prometheus_alerts = alerts_response.json()["data"]["alerts"]
-            logger.info(f"Prometheus Alerts: {prometheus_alerts}")
-            for target_label, target_msg, target_states, target_severity in [
-                (
-                    constants.ALERT_CEPHOSDSLOWOPS,
-                    "OSD requests are taking too long to process.",
-                    ["firing"],
-                    "warning",
-                )
-            ]:
-                try:
-                    prometheus.check_alert_list(
-                        label=target_label,
-                        msg=target_msg,
-                        alerts=prometheus_alerts,
-                        states=target_states,
-                        severity=target_severity,
-                        ignore_more_occurences=True,
-                    )
-                    self.test_pass = True
-                except AssertionError:
-                    logger.info(
-                        "workload storage utilization job did not finish\n"
-                        f"current utilization {round(get_percent_used_capacity(), 1)}p\n"
-                        f"time passed since start workload: {round(time.perf_counter() - self.start_workload_time)}s\n"
-                        f"timeout = {round(self.timeout_sec)}s"
-                    )
-            if self.test_pass:
-                break
-        else:
-            # if test got to this point, the alert was found, test PASS
-            pytest.fail(
-                f"failed to get 'CephOSDSlowOps' while workload filled up the storage to {self.full_osd_ratio} percents"
-            )
+        msg = f"no errors should be reported by fio writing on {vol_type} volume"
+        assert fio["jobs"][0]["error"] == 0, msg
